# openapi/campushire.openapi.json
{
  "components": {
    "schemas": {
      "ActivationStage": {
        "properties": {
          "estimated_minutes": {
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "key": {
            "enum": [
              "account_activated",
              "profile_minimum",
              "target_role",
              "resume_reviewed",
              "opportunities_unlocked",
              "first_application"
            ],
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "status": {
            "enum": [
              "complete",
              "current",
              "upcoming"
            ],
            "title": "Status",
            "type": "string"
          },
          "unlocks": {
            "title": "Unlocks",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "status",
          "href",
          "estimated_minutes",
          "unlocks"
        ],
        "title": "ActivationStage",
        "type": "object"
      },
...
      "DashboardEvidence": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "status": {
            "enum": [
              "verified",
              "pending",
              "review"
            ],
            "title": "Status",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value",
          "status"
        ],
        "title": "DashboardEvidence",
        "type": "object"
      },
      "DashboardOpportunity": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "eligibility": {
            "title": "Eligibility",
            "type": "string"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "match": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "role",
          "location",
          "eligibility",
          "match",
          "href"
        ],
        "title": "DashboardOpportunity",
        "type": "object"
      },
      "DashboardResponse": {
        "properties": {
          "activation": {
            "items": {
              "$ref": "#/components/schemas/ActivationStage"
            },
            "title": "Activation",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/DashboardEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "next_action": {
            "$ref": "#/components/schemas/NextAction"
          },
          "opportunities": {
            "items": {
              "$ref": "#/components/schemas/DashboardOpportunity"
            },
            "title": "Opportunities",
            "type": "array"
          },
          "readiness": {
            "title": "Readiness",
            "type": "integer"
          },
          "roadmap": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoadmapResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "enum": [
              "ready",
              "incomplete",
              "processing",
              "manual-review",
              "ai-unavailable"
            ],
            "title": "State",
            "type": "string"
          },
          "student_name": {
            "title": "Student Name",
            "type": "string"
          },
          "unread_notifications": {
            "title": "Unread Notifications",
            "type": "integer"
          }
        },
        "required": [
          "student_name",
          "readiness",
          "state",
          "next_action",
          "activation",
          "evidence",
          "opportunities",
          "roadmap",
          "unread_notifications"
        ],
        "title": "DashboardResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NextAction": {
        "properties": {
          "completion_criteria": {
            "title": "Completion Criteria",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "estimated_minutes": {
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "source_facts": {
            "items": {
              "type": "string"
            },
            "title": "Source Facts",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unlocks": {
            "title": "Unlocks",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "description",
          "reason",
          "href",
          "policy_version",
          "source_facts",
          "estimated_minutes",
          "unlocks",
          "completion_criteria"
        ],
        "title": "NextAction",
        "type": "object"
      },
      "NotificationCreate": {
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Body",
            "type": "string"
          },
          "deep_link": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Deep Link",
            "type": "string"
          },
          "event_key": {
            "maxLength": 180,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.:-]+$",
            "title": "Event Key",
            "type": "string"
          },
          "recipient_user_id": {
            "format": "uuid",
            "title": "Recipient User Id",
            "type": "string"
          },
          "title": {
            "maxLength": 180,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "recipient_user_id",
          "event_key",
          "title",
          "body",
          "deep_link"
        ],
        "title": "NotificationCreate",
        "type": "object"
      },
      "NotificationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread_count"
        ],
        "title": "NotificationPage",
        "type": "object"
      },
      "NotificationResponse": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deep_link": {
            "title": "Deep Link",
            "type": "string"
          },
          "event_key": {
            "title": "Event Key",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_key",
          "title",
          "body",
          "deep_link",
          "read_at",
          "created_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
...
      "RoadmapAvailabilityResponse": {
        "properties": {
          "guidance_provider_status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Guidance Provider Status",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "enum": [
              "available",
              "no_target_role",
              "no_approved_template",
              "institution_restriction"
            ],
            "title": "Status",
            "type": "string"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/RoadmapTemplateResponse"
            },
            "title": "Templates",
            "type": "array"
          }
        },
        "required": [
          "status",
          "reason",
          "guidance_provider_status",
          "templates"
        ],
        "title": "RoadmapAvailabilityResponse",
        "type": "object"
      },
      "RoadmapNodeResponse": {
        "properties": {
          "completion": {
            "title": "Completion",
            "type": "string"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "state": {
            "enum": [
              "completed",
              "next",
              "locked"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "completion",
          "prerequisites",
          "state",
          "evidence"
        ],
        "title": "RoadmapNodeResponse",
        "type": "object"
      },
      "RoadmapProgressUpdate": {
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "evidence_label": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Label"
          },
          "evidence_reference": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Reference"
          }
        },
        "required": [
          "completed"
        ],
        "title": "RoadmapProgressUpdate",
        "type": "object"
      },
      "RoadmapResponse": {
        "properties": {
          "completed_count": {
            "title": "Completed Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/RoadmapNodeResponse"
            },
            "title": "Nodes",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "template_id",
          "slug",
          "title",
          "version",
          "summary",
          "completed_count",
          "nodes"
        ],
        "title": "RoadmapResponse",
        "type": "object"
      },
      "RoadmapSelection": {
        "properties": {
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "template_id"
        ],
        "title": "RoadmapSelection",
        "type": "object"
      },
      "RoadmapTemplateResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "node_count": {
            "title": "Node Count",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "version",
          "summary",
          "node_count"
        ],
        "title": "RoadmapTemplateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CampusHire AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/notifications": {
      "post": {
        "operationId": "create_notification_api_v1_admin_notifications_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Notification",
        "tags": [
          "placement communication"
        ]
      }
    },
...
    "/api/v1/dashboard": {
      "get": {
        "operationId": "read_dashboard_api_v1_dashboard_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Dashboard",
        "tags": [
          "readiness and communication"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "read_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Notifications",
        "tags": [
          "readiness and communication"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Notification",
        "tags": [
          "readiness and communication"
        ]
      }
    },
...
    "/api/v1/roadmaps/availability": {
      "get": {
        "operationId": "read_roadmap_availability_api_v1_roadmaps_availability_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapAvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Roadmap Availability",
        "tags": [
          "readiness and communication"
        ]
      }
    },
    "/api/v1/roadmaps/current": {
      "get": {
        "operationId": "read_current_roadmap_api_v1_roadmaps_current_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RoadmapResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Read Current Roadmap Api V1 Roadmaps Current Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Current Roadmap",
        "tags": [
          "readiness and communication"
        ]
      }
    },
    "/api/v1/roadmaps/nodes/{node_key}": {
      "post": {
        "operationId": "record_roadmap_progress_api_v1_roadmaps_nodes__node_key__post",
        "parameters": [
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapProgressUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Roadmap Progress",
        "tags": [
          "readiness and communication"
        ]
      }
    },
    "/api/v1/roadmaps/select": {
      "post": {
        "operationId": "choose_roadmap_api_v1_roadmaps_select_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose Roadmap",
        "tags": [
          "readiness and communication"
        ]
      }
    },
    "/api/v1/roadmaps/templates": {
      "get": {
        "operationId": "read_roadmap_templates_api_v1_roadmaps_templates_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "campushire_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campushire Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoadmapTemplateResponse"
                  },
                  "title": "Response Read Roadmap Templates Api V1 Roadmaps Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Roadmap Templates",
        "tags": [
          "readiness and communication"
        ]
      }
    }
  }
}